feat: restructuresd to seperate modules partially done
